Make rest of Gift fields non-final and annotate them
The intention has now changed!
The User (service) is responsible for creating a Gift, but not for
persisting data about it!

Therefore a lot of the fields are @Transient
We will only persist Gift IDs and the state those Gifts are in (which
Sell/Receive lists do they correspond to)

Non-final fields are required when the Gift object is constructed
dynamically, as in JPA (see the no-args constructor).