Add Controller logic for the routes
On /register we get a UserBasicInfo and return a User written in the DB
The GET for /users and /user/{id} do what one would expect from REST
(get all collection, get a specific item in collection). Note that
this is in UserPublicInfo form.

There is also a /user_authenticated/{id} route for what the user would
see when authenticated as that id